+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Database schema: add uniqueness constraint to Check.code
 - Database schema: add Ping.kind field
 - Database schema: remove Ping.start and Ping.fail fields
+- Add "Email Settings..." dialog and "Subject Must Contain" setting
 
 
 ## 1.4.0 - 2018-12-25